changes done to structure, stream endpoint added, some files deleted that are not used
