+++ src/pages/api/video.ts
@@ -1,8 +1,6 @@
 import { NextApiRequest, NextApiResponse } from "next";
 import { downloadTorrent } from "../../server/torrent/downloadTorrent";
-import ffmpeg from 'fluent-ffmpeg';
-import fs from "fs";
-
+import { createStream }from "../../server/torrent/createStream";
 interface torrentDataInter { 
 	url: string,
 	hash: string,
@@ -23,54 +21,20 @@ const createMagnetLink = (torrents: torrentDataInter[], movieTitle: string): str
 	return magnetLink;
 }
 
-const startStream = (req: NextApiRequest, res: NextApiResponse, moviePath: string) => {
-	const range = req.headers.range
-	if (!range) {
-		res.status(400).send('Requires Range header');
-	} else {
-		const videoPath = `${moviePath}`;
-		const isMp4 = videoPath.endsWith('mp4');
-		const videoSize = fs.statSync(`${moviePath}`).size;
-		const CHUNK_SIZE = 10 ** 4; 
-		const start = Number(range.replace(/\D/g, ""));
-		const end = Math.min(start + CHUNK_SIZE, videoSize - 1); // 
-		const contentLength = end - start + 1;
-		const headers = isMp4
-			? {
-			'Content-Range': `bytes ${start}-${end}/${videoSize}`,
-			'Accept-Ranges': 'bytes',
-			'Content-Length': contentLength,
-			'Content-Type': 'video/mp4',
-			}
-			: {
-			'Content-Range': `bytes ${start}-${end}/${videoSize}`,
-			'Accept-Ranges': 'bytes',
-			'Content-Type': 'video/webm',
-			};
-		res.writeHead(206, headers);
-		const videoStream = fs.createReadStream(videoPath, { start, end });
-		if (isMp4) {
-			videoStream.pipe(res);
-		} else {
-		  ffmpeg(videoStream)
-			.format('webm')
-			.on('error', () => {})
-			.pipe(res);
-		}
-	}
-}
-
 export default async function streamVideo(req: NextApiRequest, res: NextApiResponse) {
+	console.log('in api endpoint')
 	if(req.method === 'POST') {
-		const data = req.body;
+		const data = JSON.parse(req.body);
 		const id: number = data.id;
-		const imdbCode: string = data.imdbCode;
-		const movieTitle: string = data.titleLong;
+		const imdbCode: string = data.imdb_code;
+		const movieTitle: string = data.title_long;
 		const torrents: torrentDataInter[] = data.torrents;
 		const uri: string = createMagnetLink(torrents, movieTitle);
 		const moviePath: string = `../../movies/${imdbCode}`; // THIS IS WRONG have to fix when player gets ready
 		await downloadTorrent(uri, imdbCode); // maybe pass imdb code here as well for the path of downloaded file.
-		startStream(req, res, moviePath);
+		// if there is a issue when starting to view the movie it might have something to do with that it hasnt started downloading yet
+		// before clicking play because i removed now the createStream from here.
+	} else {
+		res.status(404).json({ message: 'Type of request invalid, please do a POST request'});
 	}
-	// res.status(200).json({ name: 'John Doe' })
 }
