+++ src/pages/home/[movieId].tsx
@@ -1,6 +1,6 @@
 import { useRouter } from 'next/router';
 import { useEffect } from 'react';
-import { Container, Image } from 'react-bootstrap';
+import { Container, Image, Button } from 'react-bootstrap';
 import { useSelector } from 'react-redux';
 import { RootReducer } from '../../types/appTypes';
 import { useState } from 'react';
@@ -13,6 +13,7 @@ const MoviePage = () => {
 	);
 	const movieId = router.query.movieId;
 	const [movie, setMovie] = useState<Movie>();
+	const [isLoading, setLoading] = useState(false);
 
 	let i = 0;
 	useEffect(() => {
@@ -25,6 +26,16 @@ const MoviePage = () => {
 		}
 	}, [moviesReducer]);
 
+	const handleClick = () => {
+		setLoading(true);
+		streamMovie();
+	};
+	const streamMovie = () => {
+		const response = fetch('/api/video/', {
+			method: 'POST',
+			body: JSON.stringify(movie),
+		});
+	};
 	if (!movie?.id) {
 		return <></>;
 	} else {
@@ -46,16 +57,46 @@ const MoviePage = () => {
 							}}
 							fluid
 						>
+							<Button
+								variant="primary"
+								hidden={isLoading}
+								onClick={!isLoading ? handleClick : null}
+							>
+								play
+							</Button>
 							<Container className="overflow-hidden p-0">
-								<Image
-									className="w-100"
-									style={{
-										objectFit: 'cover',
-										minHeight: '720px',
-										maxHeight: '60vh',
-									}}
-									src={movie.background_image_original}
-								/>
+								{!isLoading && (
+									<Image
+										className="w-100"
+										style={{
+											objectFit: 'cover',
+											minHeight: '720px',
+											maxHeight: '60vh',
+										}}
+										src={movie.background_image_original}
+									/>
+								)}
+								{
+									isLoading
+									// <ReactPlayer
+									// 	url={props.url}
+									// 	controls={true}
+									// 	config={{
+									// 		file: {
+									// 			tracks: props.subtitles || [],
+									// 			attributes: {
+									// 				controlsList: 'nodownload',
+									// 			},
+									// 		},
+									// 	}}
+									// 	className={classes.reactPlayer}
+									// 	playing={false}
+									// 	width="100%"
+									// 	height="100%"
+									// 	onStart={props.onStart}
+									// 	light={props.thumbnail}
+									// />
+								}
 							</Container>
 						</Container>
 					</Container>
