--- src/server/torrent/testObject.ts
@@ -1,21 +1,0 @@
-const haha = {
-	url: "https://yts.mx/torrent/download/BF4881C718D730DE5A35A598A640B07F9FF91071",
-	hash: "BF4881C718D730DE5A35A598A640B07F9FF91071",
-	quality: "720p",
-	type: "web",
-	seeds: 0,
-	peers: 0,
-	size: "939.65 MB",
-	size_bytes: 985294438,
-	date_uploaded: "2022-10-28 06:45:28",
-	date_uploaded_unix: 1666932328
-}
-
-const torrents = [haha];
-
-const testObj = {
-	id: 46185,
-	imdbCode: "tt13603778",
-	movieTitle: "American Murderer (2022)",
-	torrents: haha
-}
